Move JsonType and subclasses to their own library
